+++ src/com/azale/engine/gfx/threedimensional/Object3D.java
@@ -2,6 +2,6 @@ package com.azale.engine.gfx.threedimensional;
 
 public class Object3D {
 
-    public int x, y, z;
+    public double x, y, z;
 
 }
